airtable refactor
