+++ src/lib/airtable/shared.ts
@@ -1,26 +1,6 @@
-import Airtable, { type FieldSet, type Record } from "airtable";
-import type { User, Job } from "./types";
-import { UserSchema, JobSchema } from "./types";
-import { z } from "zod";
-
-if (!process.env.AIRTABLE_API_KEY) {
-  throw new Error("Missing AIRTABLE_API_KEY");
-}
-
-if (!process.env.AIRTABLE_BASE_ID) {
-  throw new Error("Missing AIRTABLE_BASE_ID");
-}
-
-// Initialize Airtable
-const airtable = new Airtable({
-  apiKey: process.env.AIRTABLE_API_KEY,
-});
-
-const base = airtable.base(process.env.AIRTABLE_BASE_ID);
-
-// Table references
-const jobsTable = base("Form");
-const usersTable = base("Users");
+import type { FieldSet, Record } from "airtable";
+import { usersTable, jobsTable } from ".";
+import { JobSchema, UserSchema, type Job, type User } from "../types";
 
 // Helper functions
 export async function getById<T extends "job" | "user">(
@@ -102,59 +82,5 @@ export async function updateBySlackId<T extends "job" | "user">(
   } catch (error) {
     console.error("Error updating record:", error);
     return false;
-  }
-}
-
-// Search helper
-export async function searchJobs(query?: string) {
-  try {
-    const records = await jobsTable
-      .select({
-        filterByFormula: query
-          ? `OR(SEARCH("${query}", LOWER({slack_id})), SEARCH("${query}", LOWER({ysws})))`
-          : "",
-      })
-      .firstPage();
-
-    return records
-      .map((record) => {
-        const parsed = JobSchema.safeParse(record.fields);
-        if (!parsed.success) {
-          console.error("Failed to parse job record:", parsed.error.message);
-          return null;
-        }
-        return { ...parsed.data, id: record.id };
-      })
-      .filter(Boolean) as (Job & { id: string })[];
-  } catch (error) {
-    console.error("Error searching jobs:", error);
-    return [];
-  }
-}
-
-// Pagination helper
-export async function getUsers(page = 1, pageSize = 10) {
-  try {
-    const offset = (page - 1) * pageSize;
-    const records = await usersTable
-      .select({
-        pageSize,
-        offset,
-      })
-      .firstPage();
-
-    return records
-      .map((record) => {
-        const parsed = UserSchema.safeParse(record.fields);
-        if (!parsed.success) {
-          console.error("Failed to parse user record:", parsed.error);
-          return null;
-        }
-        return { ...parsed.data, id: record.id };
-      })
-      .filter(Boolean) as (User & { id: string })[];
-  } catch (error) {
-    console.error("Error fetching users:", error);
-    return [];
   }
 }
